Add missing .ConfigureAwait(false) everywhere
Also set the CA2007 severity as warning which will end up as an error since `TreatWarningsAsErrors` is enabled. (But disable them in the test project where they are not required.)

See also [Overview of .NET source code analysis][1] for more information about the .NET analyzers and how to configure them.

[1]: https://docs.microsoft.com/en-us/dotnet/fundamentals/code-analysis/overview